change coordinate data types

+++ migrations/001.do.create_tables.sql
@@ -11,8 +11,8 @@ CREATE TABLE IF NOT EXISTS restaurants (
     phone_link text NOT NULL,
     rest_image text,
     link text NOT NULL,
-    lat integer NOT NULL,
-    lng integer NOT NULL
+    lat float NOT NULL,
+    lng float NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS nightlife (
@@ -28,6 +28,6 @@ CREATE TABLE IF NOT EXISTS nightlife (
     adult boolean NOT NULL,
     area text NOT NULL,
     club_image text,
-    lat integer NOT NULL,
-    lng integer NOT NULL
+    lat float NOT NULL,
+    lng float NOT NULL
 );
